move `RegistryQueryer` to a dedicated mod
